requirements.txt
add-service==0.1.1
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==22.1.0
autopep8==1.7.0
blinker==1.4
certifi==2022.9.24
chardet==4.0.0
charset-normalizer==2.1.1
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
discord.py==2.0.1
distro==1.7.0
distro-info===1.1build1
frozenlist==1.3.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
multidict==6.0.2
netifaces==0.11.0
numpy==1.23.3
oauthlib==3.2.0
Pillow==9.2.0
py-cord==2.2.2
pycodestyle==2.9.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
PyYAML==5.4.1
requests==2.28.1
SecretStorage==3.3.1
six==1.16.0
toml==0.10.2
torch==1.12.1+cpu
torchaudio==0.12.1+cpu
torchvision==0.13.1+cpu
typing_extensions==4.4.0
ubuntu-advantage-tools==27.11.2
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.6
urllib3==1.26.12
wadllib==1.3.6
yarl==1.8.1
zipp==1.0.0
